style: Add Header to all pages (#62)
style: Add Header to all pages (#62)

What changed?
===================

This commit adds a header containing the nav bar and makes sure all pages display this.

Why are we doing this?
===================

To clean up the code and allow the header to contain the nav bar. This lets users see the nav bar on all pages allowing for easy navigation throughout the site.